extract/requirements.playwright.txt
scrapy-playwright==0.0.21
